% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmprior.R
\name{nmprior}
\alias{nmprior}
\title{Render character vector to paste into NM control stream}
\usage{
nmprior(
  model_name,
  THETA_VALUE,
  THETA_PRECISION,
  OMEGA_VALUE,
  OMEGA_TYPE = c("OMEGA2", "VAR", "CV", "SE", "logN_CV"),
  OMEGA_PRECISION,
  PRECISION = c("RSE", "SE", "CI90", "CI95"),
  .message = TRUE
)
}
\arguments{
\item{model_name}{Name and reference of the model}

\item{THETA_VALUE}{a named list of the values of theta}

\item{THETA_PRECISION}{a list of precision on theta}

\item{OMEGA_VALUE}{a named list of the values of omega}

\item{OMEGA_TYPE}{how was inter-individual variabilty reported ?}

\item{OMEGA_PRECISION}{a list of precision on theta}

\item{PRECISION}{is precision reported as SE or RSE}

\item{.message}{display an additional reminder about other blocks.}
}
\value{
dummy result
}
\description{
Render character vector to paste into NM control stream
}
