% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmprior.R
\name{nmprior}
\alias{nmprior}
\title{Render character vector to paste into NM control stream}
\usage{
nmprior(x)
}
\arguments{
\item{x}{dummy input}
}
\value{
dummy result
}
\description{
Render character vector to paste into NM control stream
}
